move components, fix but where search was returning 404 when route was not accessed
